requirements.txt
# ============================================
# Awaxen Backend - Python Dependencies
# ============================================

# Flask Core
Flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
flasgger>=0.9.7

# Database
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.0

# Authentication
PyJWT[crypto]>=2.8.0

# HTTP Client
requests>=2.31.0
urllib3>=2.0.0

# WebSocket & Real-time
flask-socketio>=5.3.0
gevent>=23.9.0
gevent-websocket>=0.10.1

# MQTT
paho-mqtt>=2.0.0

# Celery (Async Tasks)
celery[redis]>=5.3.0
redis>=5.0.0
flower>=2.0.0

# Encryption (Token Security)
cryptography>=41.0.0

# Environment
python-dotenv>=1.0.0

# Utilities
python-slugify>=8.0.0

# Production Server
gunicorn>=21.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-flask>=1.3.0
faker>=22.0.0

# Code Quality
black>=23.0.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.7.0

# Type Stubs
types-requests>=2.31.0
types-redis>=4.6.0
